core(metrics): move TTCI to computed artifact
