+++ lighthouse-core/gather/computed/metrics/metric.js
@@ -1,0 +1,59 @@
+/**
+ * @license Copyright 2018 Google Inc. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ */
+'use strict';
+
+const ComputedArtifact = require('../computed-artifact');
+
+class MetricArtifact extends ComputedArtifact {
+  /** @type {string} */
+  get name() {
+    throw new Error('Unimplemented');
+  }
+
+  /**
+   * @param {LH.Gatherer.Artifact.MetricComputationData} data
+   * @param {Object} artifacts
+   * @return {Promise<LH.Gatherer.Artifact.LanternMetric>}
+   */
+  computeSimulatedMetric(data, artifacts) {
+    return artifacts[`requestLantern${this.name}`](data);
+  }
+
+  /**
+   * @param {LH.Gatherer.Artifact.MetricComputationData} data
+   * @param {Object} artifacts
+   * @return {Promise<LH.Gatherer.Artifact.Metric>}
+   */
+  computeObservedMetric(data, artifacts) { // eslint-disable-line no-unused-vars
+    throw new Error('Unimplemented');
+  }
+
+  /**
+   * @param {LH.Gatherer.Artifact.MetricComputationData} data
+   * @param {Object} artifacts
+   * @return {Promise<LH.Gatherer.Artifact.LanternMetric|LH.Gatherer.Artifact.Metric>}
+   */
+  async compute_(data, artifacts) {
+    const {trace, devtoolsLog, settings} = data;
+    const augmentedData = Object.assign({
+      networkRecords: await artifacts.requestNetworkRecords(devtoolsLog),
+      traceOfTab: await artifacts.requestTraceOfTab(trace),
+    }, data);
+
+    const throttlingMethod = data.settings.throttlingMethod;
+    switch (throttlingMethod) {
+      case 'simulate':
+        return this.computeSimulatedMetric(augmentedData, artifacts);
+      case 'provided':
+      case 'devtools':
+        return this.computeObservedMetric(augmentedData, artifacts);
+      default:
+        throw new TypeError(`Unrecognized throttling method: ${throttlingMethod}`);
+    }
+  }
+}
+
+module.exports = MetricArtifact;
