+++ lighthouse-core/test/audits/consistently-interactive-test.js
@@ -16,16 +16,6 @@ const acceptableDevToolsLog = require('../fixtures/traces/progressive-app-m60.de
 const redirectTrace = require('../fixtures/traces/site-with-redirect.json');
 const redirectDevToolsLog = require('../fixtures/traces/site-with-redirect.devtools.log.json');
 
-function generateNetworkRecords(records, navStart) {
-  const navStartInMs = navStart / 1000;
-  return records.map(item => {
-    return {
-      finished: typeof item.finished === 'undefined' ? true : item.finished,
-      startTime: (item.start + navStartInMs) / 1000,
-      endTime: item.end === -1 ? -1 : (item.end + navStartInMs) / 1000,
-    };
-  });
-}
 
 /* eslint-env mocha */
 describe('Performance: consistently-interactive audit', () => {
@@ -39,7 +29,8 @@ describe('Performance: consistently-interactive audit', () => {
       },
     }, Runner.instantiateComputedArtifacts());
 
-    return ConsistentlyInteractive.audit(artifacts, {options}).then(output => {
+    const settings = {throttlingMethod: 'provided'};
+    return ConsistentlyInteractive.audit(artifacts, {options, settings}).then(output => {
       assert.equal(output.score, 0.99);
       assert.equal(Math.round(output.rawValue), 1582);
       assert.equal(output.displayValue, '1,580\xa0ms');
@@ -56,142 +47,11 @@ describe('Performance: consistently-interactive audit', () => {
       },
     }, Runner.instantiateComputedArtifacts());
 
-    return ConsistentlyInteractive.audit(artifacts, {options}).then(output => {
+    const settings = {throttlingMethod: 'provided'};
+    return ConsistentlyInteractive.audit(artifacts, {options, settings}).then(output => {
       assert.equal(output.score, 0.95);
       assert.equal(Math.round(output.rawValue), 2712);
       assert.equal(output.displayValue, '2,710\xa0ms');
-    });
-  });
-
-  describe('#findOverlappingQuietPeriods', () => {
-    it('should return entire range when no activity is present', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 2500 * 1000 + navigationStart;
-      const traceEnd = 10000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [];
-      const network = generateNetworkRecords([], navigationStart);
-
-      const result = ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      assert.deepEqual(result.cpuQuietPeriod, {start: 0, end: traceEnd / 1000});
-      assert.deepEqual(result.networkQuietPeriod, {start: 0, end: traceEnd / 1000});
-    });
-
-    it('should throw when trace ended too soon after FMP', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 2500 * 1000 + navigationStart;
-      const traceEnd = 5000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [];
-      const network = generateNetworkRecords([], navigationStart);
-
-      assert.throws(() => {
-        ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      }, /NO.*IDLE_PERIOD/);
-    });
-
-    it('should throw when CPU is quiet but network is not', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 2500 * 1000 + navigationStart;
-      const traceEnd = 10000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [];
-      const network = generateNetworkRecords([
-        {start: 1400, end: 1900},
-        {start: 2000, end: 9000},
-        {start: 2000, end: 8000},
-        {start: 2000, end: 8500},
-      ], navigationStart);
-
-      assert.throws(() => {
-        ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      }, /NO.*NETWORK_IDLE_PERIOD/);
-    });
-
-    it('should throw when network is quiet but CPU is not', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 2500 * 1000 + navigationStart;
-      const traceEnd = 10000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [
-        {start: 3000, end: 8000},
-      ];
-      const network = generateNetworkRecords([
-        {start: 0, end: 1900},
-      ], navigationStart);
-
-      assert.throws(() => {
-        ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      }, /NO.*CPU_IDLE_PERIOD/);
-    });
-
-    it('should handle unfinished network requests', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 2500 * 1000 + navigationStart;
-      const traceEnd = 10000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [];
-      const network = generateNetworkRecords([
-        {start: 0, end: -1, finished: false},
-        {start: 0, end: -1, finished: false},
-        {start: 0, end: -1, finished: false},
-      ], navigationStart);
-
-      assert.throws(() => {
-        ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      }, /NO.*NETWORK_IDLE_PERIOD/);
-    });
-
-    it('should find first overlapping quiet period', () => {
-      const navigationStart = 220023532;
-      const firstMeaningfulPaint = 10000 * 1000 + navigationStart;
-      const traceEnd = 45000 * 1000 + navigationStart;
-      const traceOfTab = {timestamps: {navigationStart, firstMeaningfulPaint, traceEnd}};
-
-      const cpu = [
-        // quiet period before FMP
-        {start: 9000, end: 9900},
-        {start: 11000, end: 13000},
-        // quiet period during network activity
-        {start: 18500, end: 22000},
-        {start: 23500, end: 26000},
-        // 2nd quiet period during network activity
-        {start: 31500, end: 34000},
-        // final quiet period
-      ];
-
-      const network = generateNetworkRecords([
-        // initial page load + script
-        {start: 1400, end: 1900},
-        {start: 1900, end: 9000},
-        // script requests more content
-        {start: 11500, end: 18500},
-        {start: 11500, end: 19000},
-        {start: 11500, end: 19000},
-        {start: 11500, end: 19500},
-        // quiet period during Main thread activity
-        {start: 28000, end: 32000},
-        {start: 28000, end: 32000},
-        {start: 28000, end: 35000},
-        // final quiet period
-      ], navigationStart);
-
-      const result = ConsistentlyInteractive.findOverlappingQuietPeriods(cpu, network, traceOfTab);
-      assert.deepEqual(result.cpuQuietPeriod, {
-        start: 34000 + navigationStart / 1000,
-        end: traceEnd / 1000,
-      });
-      assert.deepEqual(result.networkQuietPeriod, {
-        start: 32000 + navigationStart / 1000,
-        end: traceEnd / 1000,
-      });
-      assert.equal(result.cpuQuietPeriods.length, 3);
-      assert.equal(result.networkQuietPeriods.length, 2);
     });
   });
 });
