+++ typings/gatherer.d.ts
@@ -24,6 +24,20 @@ declare global {
     }
 
     namespace Artifact {
+      export interface MetricComputationData {
+        devtoolsLog: Array<Protocol.RawEventMessage>;
+        trace: {traceEvents: Array<TraceEvent>};
+        settings: ConfigSettings;
+
+        networkRecords?: Array<WebInspector.NetworkRequest>;
+        traceOfTab?: TraceOfTab;
+      }
+
+      export interface Metric {
+        timing: number;
+        timestamp: number;
+      }
+
       export interface LanternMetric {
         timing: number;
         optimisticEstimate: Simulation.Result;
